+++ admin/daftar_paket_param.php
@@ -37,7 +37,7 @@ $orderby = " ORDER BY id_paket asc";
 $sql = "SELECT id_paket, nama_paket, harga, lab_terkait, COUNT(id_param) FROM tbl_paket_param GROUP BY id_paket" . $queryCondition;
 $href = 'daftar_paket_param.php';
 
-$perPage = 10;
+$perPage = 7;
 $page = 1;
 if (isset($_POST['page'])) {
     $page = $_POST['page'];
@@ -102,7 +102,7 @@ if (!empty($result)) {
                                             <td><?php echo $result[$k]["harga"]; ?></td>
                                             <td><?php echo $result[$k]["COUNT(id_param)"]; ?> Item</td>
                                             <td>
-                                                <a type="button" class="btn btn-outline-danger btn-xs fas fa-trash-alt" href="delete_paket_param.php?ID=<?php echo $result[$k]["id_paket"]; ?>"></a>
+                                                <a type="button" class="btn btn-outline-danger btn-xs fas fa-trash-alt" onclick="return confirm('Yakin Hapus?')" href="delete_paket_param.php?ID=<?php echo $result[$k]["id_paket"]; ?>"> Hapus</a>
                                             </td>
                                         </tr>
                                 <?php
