+++ admin/history_pendaftaran.php
@@ -4,6 +4,7 @@ require_once("perpage.php");
 require_once("../database.php");
 $db_handle = new Koneksi();
 
+$no_rm = "";
 $nama_pendaftar = "";
 $tanggal = "";
 
@@ -13,7 +14,7 @@ if (!empty($_POST["search"])) {
     foreach ($_POST["search"] as $k => $v) {
         if (!empty($v)) {
 
-            $queryCases = array("nama_pendaftar", "tanggal");
+            $queryCases = array("no_rm", "nama_pendaftar", "tanggal");
             if (in_array($k, $queryCases)) {
                 if (!empty($queryCondition)) {
                     $queryCondition .= " AND ";
@@ -23,6 +24,10 @@ if (!empty($_POST["search"])) {
                 }
             }
             switch ($k) {
+                case "no_rm":
+                    $nama_pendaftar = $v;
+                    $queryCondition .= "rm.no_rm LIKE '" . $v . "%'";
+                    break;
                 case "nama_pendaftar":
                     $nama_pendaftar = $v;
                     $queryCondition .= "rm.nama LIKE '" . $v . "%'";
@@ -39,7 +44,7 @@ $orderby = " ORDER BY p.id_pendaftaran desc";
 $sql = "SELECT p.id_pendaftaran, rm.no_rm, rm.nama, p.tgl_daftar FROM tbl_pendaftaran as p JOIN tbl_rm as rm ON p.no_rm = rm.no_rm" . $queryCondition . " GROUP BY p.id_pendaftaran";
 $href = 'history_pendaftaran.php';
 
-$perPage = 10;
+$perPage = 7;
 $page = 1;
 if (isset($_POST['page'])) {
     $page = $_POST['page'];
@@ -62,7 +67,7 @@ if (!empty($result)) {
         <div class="container-fluid">
             <div class="row mb-2">
                 <div class="col-sm-6">
-                    <h1>Input Hasil LAB</h1>
+                    
                 </div>
             </div>
         </div><!-- /.container-fluid -->
@@ -72,28 +77,35 @@ if (!empty($result)) {
     <section class="content">
         <div class="card">
             <div class="card-header">
-                <a class="btn btn-primary" href="daftar_pendaftaran.php">+ Tambah Pendaftaran</a>
+                <h3 class="card-title">Input Hasil LAB</h3>
+                <div class="card-tools">
+                    <a class="btn btn-primary btn-sm" href="daftar_pendaftaran.php">+ Tambah Pendaftaran</a>
+                </div>
             </div>
             <!-- /.card-header -->
             <div class="card-body">
                 <form autocomplete="off" name="frmSearch" method="post" action="history_pendaftaran.php">
                     <div class="search-box">
-                        <p>
-                            <input type="text" placeholder="Nama Pendaftar" name="search[nama_pendaftar]" class="" value="<?php echo $nama_pendaftar; ?>" />
-                            <input type="date" name="search[tanggal]" class="" value="<?php echo $tanggal; ?>" />
-                            <input type="submit" name="go" class="btnSearch" value="Search">
-                            <input type="reset" class="btnSearch" value="Reset" onclick="window.location='history_pendaftaran.php'">
-                        </p>
+                        <table>
+                            <tr>
+                                <td><input type="text" placeholder="No RM" name="search[no_rm]" class="form-control" value="<?php echo $no_rm; ?>" /></td>
+                                <td><input type="text" placeholder="Nama Pendaftar" name="search[nama_pendaftar]" class="form-control" value="<?php echo $nama_pendaftar; ?>" /></td>
+                                <td><input type="date" name="search[tanggal]" class="form-control" value="<?php echo $tanggal; ?>" /></td>
+                                <td><input type="submit" name="go" class="btnSearch btn btn-outline-success" value="Search"></td>
+                                <td><input type="reset" class="btnSearch btn btn-outline-secondary" value="Reset" onclick="window.location='history_pendaftaran.php'"></td>
+                            </tr>
+                        </table>
+                        <br />
                     </div>
 
                     <table class="table table-bordered">
                         <thead>
-                            <tr>
+                            <tr align="center">
                                 <th>No REG</th>
                                 <th>No RM Pendaftar</th>
                                 <th>Nama Pendaftar</th>
                                 <th>Tanggal Mendaftar</th>
-                                <th>Action</th>
+                                <th colspan="2">Action</th>
                             </tr>
                         </thead>
                         <tbody>
@@ -107,9 +119,12 @@ if (!empty($result)) {
                                             <td><?php echo $result[$k]["no_rm"]; ?></td>
                                             <td><?php echo $result[$k]["nama"]; ?></td>
                                             <td><?php echo $result[$k]["tgl_daftar"]; ?></td>
-                                            <td>
+                                            <td align="center">
                                                 <a type="button" class="btn btn-outline-success btn-xs fas fa-check-square" href="input_hasil_lab.php?id=<?php echo $result[$k]["id_pendaftaran"]; ?>"> Pilih</a>
-                                                <a type="button" class="btn btn-outline-danger btn-xs fas fa-trash-alt" href="delete_history_pendaftaran.php?id=<?php echo $result[$k]["id_pendaftaran"]; ?>"> Hapus</a>
+                                            </td>
+                                            <td align="center">
+                                                <a type="button" class="btn btn-outline-primary btn-xs fas fa-edit" href="edit_history_pendaftaran.php?id=<?php echo $result[$k]["id_pendaftaran"]; ?>"> Edit</a>
+                                                <a type="button" class="btn btn-outline-danger btn-xs fas fa-trash-alt" onclick="return confirm('Yakin Hapus?')" href="delete_history_pendaftaran.php?id=<?php echo $result[$k]["id_pendaftaran"]; ?>"> Hapus</a>
                                             </td>
                                         </tr>
                                 <?php
